requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-xdist==3.3.1
allure-pytest==2.13.2
python-dotenv==1.0.0
faker==20.1.0